# Improve LibraryManager Dialog rendering