TableCellRender is designed to be reusable.
Te old implementation
created an instance, every time the mouse was over the line, This can lead to rendering problems, and low responsiveness

### All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](https://github.com/arduino/Arduino/pulls?q=) for the same update/change?

### Tests

I made a small benchmark with the results:

First run:
- OLD: 423813 ms
- New: 286341 ms

Second run:
- OLD: 442275 ms
- New: 277634 ms

Test Class..

```
package cc.arduino.test;

import java.awt.Rectangle;
import java.awt.Toolkit;

import javax.swing.JTable;
import javax.swing.SwingUtilities;

public class BenchmarkRenderer {
  static long benchmarkTable(final JTable list, String msg) {
    Runnable doScroll = new Runnable() {
      public void run() {
        int index = list.getSelectedRow();
        list.changeSelection(index + 1, 0, false, false);
        // make visible to trigger render...
        Rectangle cellRect = list.getCellRect(index + 1, 0, false);
        list.scrollRectToVisible(cellRect);
      }
    };

    Runnable doNothing = new Runnable() {
      public void run() {
      }
    };

    long startTime = System.currentTimeMillis();

    list.changeSelection(0, 0, false, false);

    /*
     * Each iteration of this loops queue a request to scroll the list on the
     * event dispatching thread (we're running on the "main" thread) and then
     * waits for all of the paint operations caused by the scroll to finish.
     * Finally we force any output that has been buffered by the AWT
     * implementation to be flushed (with Toolkit.sync()). Then overhead
     * introduced here is the same for each list so the elapsed time numbers can
     * be safely compared.
     */
    for (int i = 0; i < list.getModel().getRowCount(); i++) {
      try {
        SwingUtilities.invokeLater(doScroll);
        Thread.yield();
        SwingUtilities.invokeAndWait(doNothing);
        Toolkit.getDefaultToolkit().sync();
      } catch (Exception e) {
        e.printStackTrace();
      }
    }

    long endTime = System.currentTimeMillis();
    long time =  endTime - startTime;
    System.out.println(">>" + time);
    return time;
  }

  public static void main(String[] args) throws Exception {
   
  }
}
```