@ricardojlrufino 2020-05-25
Strange, this commit should not be that behavior.
This had happened to me some of these when I was implementing the startup improvements and I found the error: #10228 , you have tested using #10270 ?