@cmaglie 2020-05-25
mmm no just tested again, I've checkout out this PR from git (so only this commit).
On the `master` branch the bug is not present, so it has to be something with this patch but, admiteddly, I can't figure out what...